chore(utils): move fetchItems from Shared to Admin
